feat: ADD LOGIN POPUP, ADD PROFILE IN HEADER
