+++ src/utils/supabase/middleware.ts
@@ -39,7 +39,7 @@ export async function updateSession(request: NextRequest) {
     data: { user },
   } = await supabase.auth.getUser();
 
-  console.log(user);
+  // console.log(user);
 
   // if (
   //   !user &&
